test: guarantee EVENT_STATE_CHANGED const and isolate caps probe for full-suite runs
